Fixed imports after move of handleError.ts
