+++ components/DeleteModal.tsx
@@ -6,7 +6,7 @@ import { Button } from './Button';
 import { SvgCheckMarkIcon } from '~/icons/CheckMarkIcon';
 import { SvgDeclineIcon } from '~/icons/DeclineIcon';
 import { deleteGift } from '~/utils/giftRequests';
-import { handleGiftError } from '~/frontend/handleError';
+import { handleGiftError } from '~/utils/handleError';
 
 type DeleteModal = {
   gift: Gift;
